Implement data model adjustments, incremental MCP and incremental EDS
