+++ istio/src/main/java/com/alibaba/nacos/istio/common/EventProcessor.java
@@ -20,7 +20,6 @@ import com.alibaba.nacos.istio.mcp.NacosMcpService;
 import com.alibaba.nacos.istio.misc.Loggers;
 import com.alibaba.nacos.istio.model.DeltaResources;
 import com.alibaba.nacos.istio.util.IstioExecutor;
-import com.alibaba.nacos.istio.xds.NacosDeltaService;
 import com.alibaba.nacos.istio.xds.NacosXdsService;
 import com.alibaba.nacos.sys.utils.ApplicationUtils;
 import org.springframework.context.ApplicationListener;
@@ -47,8 +46,6 @@ public class EventProcessor implements ApplicationListener<ContextRefreshedEvent
     private NacosMcpService nacosMcpService;
 
     private NacosXdsService nacosXdsService;
-    
-    private NacosDeltaService nacosDeltaService;
 
     private NacosResourceManager resourceManager;
 
@@ -121,7 +118,7 @@ public class EventProcessor implements ApplicationListener<ContextRefreshedEvent
     }
 
     private boolean hasClientConnection() {
-        return nacosMcpService.hasClientConnection() || nacosXdsService.hasClientConnection() || nacosDeltaService.hasClientConnection();
+        return nacosMcpService.hasClientConnection() || nacosXdsService.hasClientConnection() || nacosXdsService.hasDeltaClientConnection();
     }
 
     private boolean needNewTask(boolean hasNewEvent, Future<Void> task) {
@@ -140,13 +137,13 @@ public class EventProcessor implements ApplicationListener<ContextRefreshedEvent
         public Void call() throws Exception {
             DeltaResources deltaResources = event.getDeltaResources();
             ResourceSnapshot snapshot = resourceManager.createResourceSnapshot(
-                    deltaResources.getRemovedHostName(),
+                    deltaResources.getRemovedServiceEntryName(),
                     deltaResources.getRemovedClusterName(),
                     deltaResources.getServiceChangeMap().keySet(),
                     deltaResources.getInstanceChangeMap().keySet());
             
             nacosXdsService.handleEvent(snapshot, event);
-            nacosDeltaService.handleEvent(snapshot, event);
+            nacosXdsService.handleDeltaEvent(snapshot, event);
             nacosMcpService.handleEvent(snapshot, event);
             return null;
         }
@@ -159,13 +156,10 @@ public class EventProcessor implements ApplicationListener<ContextRefreshedEvent
         if (null == nacosXdsService) {
             nacosXdsService = ApplicationUtils.getBean(NacosXdsService.class);
         }
-        if (null == nacosDeltaService) {
-            nacosDeltaService = ApplicationUtils.getBean(NacosDeltaService.class);
-        }
+        
         if (null == nacosMcpService) {
             nacosMcpService = ApplicationUtils.getBean(NacosMcpService.class);
         }
-        return Objects.nonNull(resourceManager) && Objects.nonNull(nacosMcpService)
-                && Objects.nonNull(nacosXdsService) && Objects.nonNull(nacosDeltaService);
+        return Objects.nonNull(resourceManager) && Objects.nonNull(nacosMcpService) && Objects.nonNull(nacosXdsService);
     }
 }
